requirements.txt
numpy
PyQt5
python-wordpress-xmlrpc
markdown 
requests